+++ src/main/java/org/springframework/samples/petclinic/repository/springdatajpa/TournamentRepository.java
@@ -8,16 +8,16 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface TournamentRepository extends CrudRepository<Tournament, Integer>{
 
-	@Query(value="SELECT DTYPE FROM TOURNAMENTS WHERE ID = ?1 ",nativeQuery=true)
+	@Query(value="select dtype from tournaments where id = ?1 ",nativeQuery=true)
 	String findTournamentType(int tournamentId);
 
-	@Query(value="SELECT CANODROME FROM TOURNAMENTS WHERE ID = ?1 ",nativeQuery=true)
+	@Query(value="select canodrome from tournaments where id = ?1 ",nativeQuery=true)
 	String getCanodrome(int tournamentId);
 	
-	@Query(value="SELECT CIRCUIT FROM TOURNAMENTS WHERE ID = ?1 ",nativeQuery=true)
+	@Query(value="select circuit from tournaments where id = ?1 ",nativeQuery=true)
 	String getCircuit(int tournamentId);
 	
-	@Query(value="SELECT PLACE FROM TOURNAMENTS WHERE ID = ?1 ",nativeQuery=true)
+	@Query(value="select place from tournaments where id = ?1 ",nativeQuery=true)
 	String getPlace(int tournamentId);
