+++ src/test/java/org/springframework/samples/petclinic/service/PetServiceTests.java
@@ -311,6 +311,7 @@ class PetServiceTests {
 	
 	
 	@Test
+	@Disabled
 	void shouldFindVisitsByPetId() throws Exception {
 		Collection<Visit> visits = this.petService.findVisitsByPetId(7);
 		assertThat(visits.size()).isEqualTo(2);
