+++ src/test/java/org/springframework/samples/petclinic/ui/ShowResultsNegativeUITest.java
@@ -2,12 +2,9 @@ package org.springframework.samples.petclinic.ui;
 
 import java.util.regex.Pattern;
 import java.util.concurrent.TimeUnit;
-
-import org.hamcrest.core.IsEqual;
 import org.junit.*;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import static org.junit.Assert.*;
@@ -21,7 +18,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class MostrarResultadosPositiveUITests {
+public class ShowResultsNegativeUITest {
   private WebDriver driver;
   private String baseUrl;
   private boolean acceptNextAlert = true;
@@ -29,7 +26,6 @@ public class MostrarResultadosPositiveUITests {
 
   @LocalServerPort
   private int port;
-  
   
   @BeforeEach
   public void setUp() throws Exception {
@@ -40,9 +36,9 @@ public class MostrarResultadosPositiveUITests {
     driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
   }
 
-  @Test
-  public void testMostrarResultados() throws Exception {
-    driver.get("http://localhost:" + this.port);
+  @org.junit.jupiter.api.Test
+  public void testMostrarResultadosNegative() throws Exception {
+	driver.get("http://localhost:" + this.port);
     driver.findElement(By.xpath("//div[@id='main-navbar']/ul[2]/li/a")).click();
 	driver.findElement(By.id("username")).clear();
 	driver.findElement(By.id("username")).sendKeys("admin1");
@@ -50,11 +46,10 @@ public class MostrarResultadosPositiveUITests {
 	driver.findElement(By.id("password")).clear();
 	driver.findElement(By.id("password")).sendKeys("4dm1n");
 	driver.findElement(By.xpath("//button[@type='submit']")).click();
-    driver.findElement(By.xpath("//div[@id='main-navbar']/ul/li[4]/a/span[2]")).click();
-    driver.findElement(By.xpath("//table[@id='tournamentsTable']/tbody/tr/td[4]")).click();
-    String status = driver.findElement(By.xpath("//table[@id='tournamentsTable']/tbody/tr/td[4]")).getText();
-    assertThat(status, IsEqual.equalTo("FINISHED"));
-    driver.findElement(By.linkText("First Race Test")).click();
+	driver.findElement(By.xpath("//div[@id='main-navbar']/ul/li[4]/a/span[2]")).click();
+    driver.findElement(By.xpath("//table[@id='tournamentsTable']/tbody/tr[2]/td[4]")).click();
+    String status = driver.findElement(By.xpath("//table[@id='tournamentsTable']/tbody/tr[2]/td[4]")).getText();
+    driver.findElement(By.linkText("Second Race Test")).click();
     driver.findElement(By.linkText("Show results")).click();
   }
 
@@ -100,3 +95,4 @@ public class MostrarResultadosPositiveUITests {
     }
   }
 }
+
